:sparkles: UPDATE: dep tree
