+++ e2e/react/src/api/dico.ts
@@ -1,0 +1,1 @@
+export const dico = { HELLO: `👋 Hey, hello` }
