+++ e2e/react/src/api/extra.ts
@@ -1,0 +1,5 @@
+export const extra = /* GraphQL */ `
+	extend type Query {
+		welcomeTooEveryone: String!
+	}
+`
